+++ src/app/components/tree/node-tree/node/terra-node.component.ts
@@ -66,26 +66,6 @@ export class TerraNodeComponent<D> implements OnInit
 
     }
 
-    /**
-     * @deprecated will be removed in the next major release.
-     */
-    protected onNodeDblClick(event:Event):void
-    {
-        event.stopPropagation();
-        // check if click function is set
-        if(!isNullOrUndefined(this.inputNode.onDblClick))
-        {
-            this.inputNode.onDblClick();
-        }
-
-        this.inputConfig.handleLazyLoading(this.inputNode);
-
-        if(isNullOrUndefined(this.inputNode.selectable) || this.inputNode.selectable)
-        {
-            this.inputConfig.currentSelectedNode = this.inputNode;
-        }
-    }
-
     protected handleIconClick(event:Event):void
     {
         event.stopPropagation();