+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+###UNRELEASED
+* **terra-node.interface**
+	- removed double click function
+
 <a name="3.0.0-beta.12"></a>
 # 3.0.0-beta.12 (07.11.2018)